Change path to controlgroup icon in ISM View.
I do not know why it works, but the path to the controlgroup icons
seems to be wrong in ImageCache. I added the suffix
"sweetie-verinice/png" and know the search result table is able to
display an icon for this type of group. The ISMView seems to make some
magic, since it first takes the children element icon and after that
it looks for the group icon of this child icon.

(VN-1083)