Do not enable export cvs button.
The RightsEnabledAction class does not override the *setEnable()*
method, so a client of this class can enable an action by accident when
calling this in a programmatic way.

To override this method may causes some side effects so this patch
checks the rights everytime before the setEnable method is called.

It should be part of future work to extend the RightsEnabledAction class
and make the check within this class.

(VN-1083)