Use the blocking index method for reindex via UI.
The re-index job called by the refresh button in the search view uses
the the non blocking method for indexing the verinice database, which
means, that the call returned immediately and the user thought, the
re-indexing is already finished but is not. Since the re-index job is
called in an thread (WorkspaceJob) it is no problem calling the blocking
method, the UI is not blocked anyway and the refresh button becomes only
active, when the indexing is really finished.

(VN-1082)