Replace the 3-Gram analyzer with standard german/english tokenizer
Reasons for this switch are as follows:

- matches are more predictable for the user.
- tokens which are shorter than 3 characters are also found.
- depending on the selected stemmer also composite words are found.
- phrases are found, eg. "Sehr hoch"

There exist two analysis_<language>.json files. The server uses
Local.getDefault().getLanguage() for deciding which one should be
taken.

The german analyzer does not defined a stop word list, which means,
that article, prepositions are anlyzed by elastic search and
found. This is done, because I haven't managed that the german
*keyword marker* is working as aspected. In contrast to that the
English *keyword marker* works and the English equivalent "very" is
found, without indexing all the typical english stopwords.

(VN-1082)