Remove stopwords and keyword markers.
Index does not significantly grow on large database.  It is clear, that
it does not make sense to search words like "der". But sometimes it
is. We want to be able to search for "sehr hoch", which is not possible
with the default german stop list. The keyword-token filter does not
provide a solution to this problem, since it only prevents the word to
be stemmed, so it is removed from the token stream by the
stopword-filter anyway.

(VN-1082)